Move flatpages -> pages. Create fixture for flatpages initial data. Use id instead of url for editing.
