+++ hotline/pages/templatetags/markdown.py
@@ -1,4 +1,5 @@
 from django import template
+from django.utils.html import mark_safe
 import markdown
 
 register = template.Library()
@@ -6,4 +7,4 @@ register = template.Library()
 
 @register.filter
 def md(text):
-    return markdown.markdown(text)
+    return mark_safe(markdown.markdown(text))
