+++ hotline/pages/views.py
@@ -2,70 +2,58 @@ from arcutils import will_be_deleted_with
 from django.shortcuts import render, get_object_or_404, redirect
 from django.core.exceptions import ObjectDoesNotExist
 from django.contrib import messages
-from django.contrib.sites.models import Site
 from django.contrib.flatpages.models import FlatPage
 
-from hotline.flatpages.forms import FlatterPageForm
+from hotline.pages.forms import FlatterPageForm
 
 from .perms import permissions
 
 
-def _list(request):
+def list_(request):
     """
 
     """
     pages = FlatPage.objects.all()
 
-    return render(request, "flatpages/list.html", {
+    return render(request, "pages/list.html", {
         'pages': pages,
     })
 
 
 @permissions.can_modify_page
 def create(request):
-    return _edit(request, page_url=None)
+    return _edit(request, page_id=None)
 
 
 @permissions.can_modify_page
-def edit(request, page_url):
-    print(page_url)
-    return _edit(request, page_url)
+def edit(request, page_id):
+    return _edit(request, page_id)
 
 
-def _edit(request, page_url):
-
-    site = Site.objects.first()
-    if page_url is None:
+def _edit(request, page_id):
+    if page_id is None:
         page = None
-        url = None
     else:
-        url = page_url
-        try:
-            page = FlatPage.objects.get(url=url)
-        except ObjectDoesNotExist:
-            url = "/" + str(url)
-            page = get_object_or_404(FlatPage, url=url)
+        page = get_object_or_404(FlatPage, pk=page_id)
 
     if request.method == "POST":
         form = FlatterPageForm(request.POST, instance=page)
-        form.fields['sites'].initial = [site]
         if form.is_valid():
             form.save()
             return redirect('pages-list')
 
     else:
         form = FlatterPageForm(instance=page)
-        form.fields['sites'].initial = [site]
 
-    return render(request, "flatpages/edit.html", {
+    return render(request, "pages/edit.html", {
         'form': form,
-        'page_url': url,
+        'page_id': page_id,
     })
 
 
 @permissions.can_modify_page
-def delete(request, page_url):
-    page = get_object_or_404(FlatPage, url=page_url)
+def delete(request, page_id):
+    page = get_object_or_404(FlatPage, pk=page_id)
     if request.method == "POST":
         page.delete()
         messages.success(request, "Page deleted!")
